plugin: fix TODO's for CRD and sample template
