+++ example/Makefile
@@ -13,7 +13,7 @@ endif
 
 # Run against the configured Kubernetes cluster in ~/.kube/config
 run:
-	go run ../main.go
+	go run ../main.go run
 
 # Install CRDs into a cluster
 install: kustomize
